requirements.txt
Django==5.0.8
djangorestframework==3.15.2
PyMySQL>=1.1.0  # MySQL 클라이언트 (mysqlclient 대신 사용)
python-dotenv==1.0.0
PyJWT==2.8.0
boto3==1.34.0  # S3 파일 업로드용
django-cors-headers==4.3.1
cryptography>=41.0.0  # PyMySQL 의존성
django-filter==24.3  # 필터링 기능 (Django 5.0 호환)

